requirements.txt
docker==7.0.0
boto3==1.34.0
flask==2.3.3
pyyaml==6.0.1
configparser==5.3.0
click==8.1.7
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
tenacity==8.2.3
docker-compose==1.29.2
